Accept Blocks
Make include/uninclude accept blocks.
If a block is given, include/uninclude will now yield (yield self) to
the block and undo what they did after the block is finished running.

I think this would be nicely in keeping it's how the rest of ruby
behaves for undoable methods (File.open comes to mind).

The whole code is in ruby space and it also makes the methods public,
which I think they are, anyway, in the latest rubies.

ON ALIASES
I don't know of a way to make overshaddow the original methods without
making aliases -- not if the method is to accept a block, anyway.

(Maybe it's possible in C space?)